+++ CHANGELOG.md
@@ -19,6 +19,7 @@ reference date, if available (e.g. live streams)
 - Duration field on edit media screen
 - Add media screen refuses to add inaccessible videos
 - Metadata store relays 'not found' and 'access denied' errors
+- [ELAN](https://archive.mpi.nl) `.eaf` file export for media entries
 
 ### Changed